# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/listing-ad-fee:
    patch:
      operationId: updateAdminListingAdFee
      tags: [admin]
      summary: Update the listing publication fee using an admin password session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListingAdFeeUpdate" }
      responses:
        "200":
          description: Updated listing publication fee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListingAdFee" }
        "400":
          description: Fee is outside the supported Pi range or precision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Password-authenticated admin session required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Request did not originate from this site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Listing publication fee settings are unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/session:
    post:
      operationId: createAdminSession
      tags: [admin]
      summary: Create a short-lived admin session with the configured secret password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSessionInput" }
      responses:
        "200":
          description: Admin password accepted
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HttpOnly, same-site admin session cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSessionResult" }
        "401":
          description: Password was not accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Request did not originate from this site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too many login attempts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Admin password or session signing secret is not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      operationId: deleteAdminSession
      tags: [admin]
      summary: Clear the current admin password session
      responses:
        "204":
          description: Admin session cleared
        "403":
          description: Request did not originate from this site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Live arbitration and payout counters for password-authenticated admin sessions
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOverview" }
        "401":
          description: Password-authenticated admin session required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /admin/disputes:
    get:
      operationId: listAdminDisputes
      tags: [admin]
      summary: List disputes and related escrow details for password-authenticated admin sessions
      responses:
        "200":
          description: Disputes, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminDispute" }
        "401":
          description: Password-authenticated admin session required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/disputes/{id}/analyze:
    post:
      operationId: analyzeAdminDispute
      tags: [admin]
      summary: Get an advisory Gemini analysis with an admin password session
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Advisory analysis for human review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnalysis" }
        "401":
          description: Password-authenticated admin session required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Request did not originate from this site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Gemini analysis unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /listing-ad-fee:
    get:
      operationId: getListingAdFee
      tags: [payments]
      summary: Read the current system-controlled listing publication fee
      responses:
        "200":
          description: Current fee used for new listing payment intents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListingAdFee" }
        "503":
          description: Listing publication fee settings are unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    piAppSessionCookie:
      type: apiKey
      in: cookie
      name: pi_app_session
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AdminSessionInput:
      type: object
      properties:
        password: { type: string, minLength: 1, maxLength: 512 }
      required: [password]
    AdminSessionResult:
      type: object
      properties:
        authenticated: { type: boolean, const: true }
        sessionToken: { type: string, minLength: 1 }
      required: [authenticated]
    AdminOverview:
      type: object
      properties:
        openDisputes: { type: integer, minimum: 0 }
        resolvedDisputes: { type: integer, minimum: 0 }
        disputedContracts: { type: integer, minimum: 0 }
        payoutAttention: { type: integer, minimum: 0 }
        totalEscrowPi: { type: string }
        payoutEnabled: { type: boolean }
      required: [openDisputes, resolvedDisputes, disputedContracts, payoutAttention, totalEscrowPi, payoutEnabled]
    AdminDispute:
      type: object
      properties:
        id: { type: string }
        contractId: { type: string, format: uuid }
        reason: { type: string }
        description: { type: string }
        requestedResolution: { type: string }
        resolution: { type: ["string", "null"] }
        status: { type: string, enum: [open, under_review, resolved] }
        decision: { type: ["string", "null"], enum: [release, refund, null] }
        contractTitle: { type: string }
        contractReference: { type: string }
        contractAmount: { type: string }
        contractCurrency: { type: string }
        contractStatus: { type: string }
        buyerName: { type: string }
        sellerName: { type: string }
        payoutStatus: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, contractId, reason, description, requestedResolution, resolution, status, decision, contractTitle, contractReference, contractAmount, contractCurrency, contractStatus, buyerName, sellerName, payoutStatus, createdAt, updatedAt]
    AdminAnalysis:
      type: object
      properties:
        summary: { type: string }
        recommendation: { type: string, enum: [release, refund, review] }
        rationale:
          type: array
          items: { type: string }
        evidenceGaps:
          type: array
          items: { type: string }
        confidence: { type: string, enum: [low, medium, high] }
        generatedAt: { type: string, format: date-time }
      required: [summary, recommendation, rationale, evidenceGaps, confidence, generatedAt]
...
    ListingAdFee:
      type: object
      properties:
        listingAdFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
        listingEditFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
        listingDeleteFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
      required: [listingAdFeePi, listingEditFeePi, listingDeleteFeePi]
    ListingAdFeeUpdate:
      type: object
      minProperties: 1
      properties:
        listingAdFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
        listingEditFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
        listingDeleteFeePi:
          type: number
          minimum: 0.00000001
          maximum: 1000000
          description: Pi amount with up to 8 decimal places; precision is validated by the API.
...